refactor: rename file names to have a consistent naming style
