+++ src/components/card-details/logic-card-item.tsx
