+++ src/components/index.ts
@@ -1,10 +1,10 @@
-export { default as Button } from './Button';
-export { default as CardDetails } from './card-details/card-details-render';
+export { default as Button } from './button';
+export { default as CardDetails } from './card-details/render-card-item';
 export { default as CardList } from './card-list/card-list';
-export { default as ErrorDataRetrieval } from './ErrorDataRetrieval';
-export { default as ErrorBoundary } from './ErrorBoundary';
-export { default as Footer } from './Footer';
-export { default as Header } from './Header';
-export { default as Input } from './Input';
+export { default as ErrorDataRetrieval } from './error-data-retrieval';
+export { default as ErrorBoundary } from './error-boundary';
+export { default as Footer } from './footer';
+export { default as Header } from './header';
+export { default as Input } from './input';
 export { default as Main } from './main/Main';
-export { default as Spinner } from './Spinner';
+export { default as Spinner } from './spinner';
